backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Network (macOS LibreSSL uyumlulugu)
urllib3<2.0
requests>=2.28.0,<3.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Machine Learning & NLP
sentence-transformers==2.2.2
transformers==4.35.0
faiss-cpu==1.7.4
scikit-learn==1.3.2
torch==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
# sqlite3 is built-in to Python

# Cache & LLM
diskcache>=5.6.0
ollama>=0.3.0

# API & Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.27.0

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0